api.py (breaking!): move_or_copy overhaul
Method signature of move_or_copy has changed: now the basic invocation
is move_or_copy(src, dst) instead of move_or_copy(path, new_parent,
new_name), which is rather paranoid.

Also, overwriting behavior is now more conservative, and matches that of
coreutils (and likely other implementations of) mv and cp. See the
updated docstring (especially that of the `overwrite` parameter) for
details.

The aliases `move` and `copy` are clearly affected in the same way.

Also in this commit: added to setup.py the onedrive-metadata script that
I previously forgot to commit in.