auth.py: critical bug fix: update session after token refresh
The critical line (previously only in __init__ of OneDriveOAuthClient):

    self.client.params.update({"access_token": self._access_token})

Without executing this after every refresh, although we have retrieved a
new code and saved it in self._access_token, what is used in requests is
still the old token, since self.client.params (params of the requests
Session) still holds the old token. That results in failure of all API
requests one hour after initializing the client.

Also in this commit:

* Refresh token and retry after a HTTP 401 Unauthorized;
* Not deducting a minute from expiration time anymore since it's random
  and ugly; if we really do get 401 in an edge case, just refresh and
  retry.

This commit supposedly fixes #16.